Crate oauth2_cli

Structs

Args 🔒

Sample program to use OAuth2 authentication.

Tokens 🔒

Output JSON struct containing tokens and associated metadata.

Enums

Command 🔒

CLI command.

OutputMode 🔒

Human-readable text or JSON output.

Functions

init 🔒

Performs the initialization flow to create new Tokens data from an oauth2::ClientInfo.

main 🔒

refresh 🔒

Performs the refresh flow to get new Tokens data from a refresh token.